requirements.txt
fastapi~=0.115.5
pydantic~=2.10.2
docker~=7.1.0
cookiecutter~=2.6.0
prometheus-client~=0.21.1
uvicorn~=0.32.1
requests~=2.32.3
llama-cpp-python~=0.3.2
pillow~=11.0.0
tensorboardx~=2.6.2.2
click~=8.1.7
numpy~=2.0.2
gtts~=2.5.4
outetts~=0.2.1
sympy~=1.13.3
torch~=2.4.1
manim~=0.18.1
aiohttp~=3.11.8